+++ src/normalize.ts
@@ -1,7 +1,5 @@
-
 import { isString } from './primitive';
 import { isEmail } from './string';
-
 
 /**
  * A function to capitalize the first letter of each word in a string.
@@ -10,20 +8,18 @@ import { isEmail } from './string';
  * @param {boolean} everyWords - A flag to indicate whether to capitalize every word or just the first letter of the whole string.
  * @return {string} The string with the first letter of each word capitalized.
  */
-function ucfirst(str: any, everyWords: boolean = true): any {
-  if (isString(str, true)) {
-    const newStr = str.toLowerCase();
+function ucfirst(s: string, everyWords = true): string | false {
+  if (!isString(s, true)) return false;
 
-    if (everyWords) {
-      const words = newStr.split(" ");
-      for (let i = 0; i < words.length; i++) {
-        words[i] = words[i].charAt(0).toUpperCase() + words[i].slice(1);
-      }
-      return words.join(" ");
+  const newStr = s.toLowerCase();
+  if (everyWords) {
+    const words = newStr.split(" ");
+    for (let i = 0; i < words.length; i++) {
+      words[i] = words[i].charAt(0).toUpperCase() + words[i].slice(1);
     }
-    return newStr.charAt(0).toUpperCase() + newStr.slice(1);
+    return words.join(" ");
   }
-  return str;
+  return newStr.charAt(0).toUpperCase() + newStr.slice(1);
 }
 
 /**
@@ -37,36 +33,51 @@ function ucfirst(str: any, everyWords: boolean = true): any {
  * @param {string} lastName - The last name of the user.
  * @return {string} The normalized nickname.
  */
-function normalizeNickname(nickname: any, firstName: any, lastName: any): string | false {
-  return isString(nickname, true) ? nickname.toLowerCase() : (isString(firstName, true) && isString(lastName, true)) ? createNickname(firstName, lastName) : false;
+function normalizeNickname(nickname: string, firstName: string, lastName: string): string | false {
+  return isString(nickname, true) || (isString(firstName, true) && isString(lastName, true)) ? createNickname(nickname, firstName, lastName) : false;
 }
 
 /**
  * Normalizes a first name by capitalizing the first letter of each word.
  *
- * @param {string} str - The first name to normalize.
+ * @param {string} s - The first name to normalize.
  * @return {string} The normalized first name.
  */
-function normalizeName(str: any): string {
-  return ucfirst(str, true);
-}
-
-function normalizeEmail(str: any): string | false {
-  return isEmail(str) ? str.toLowerCase() : false; 
+function normalizeName(s: string): string | false {
+  return ucfirst(s, true);
 }
 
 /**
- * Creates a nickname for a user based on the first letter of the first name and the last name.
+ * A function to normalize an email address.
  *
+ * If the string is not a valid email address, the function will return false.
+ *
+ * @param {string} str - The email address to normalize.
+ * @return {string|false} The normalized email address or false if the
+ * string is not a valid email address.
+ */
+function normalizeEmail(s: string): string | false {
+  return isEmail(s) ? s.toLowerCase() : false; 
+}
+
+/**
+ * Creates a normalized nickname for a user.
+ *
+ * If a nickname is given, the function will return that nickname with
+ * any accents and diacritics removed. If no nickname is given, the
+ * function will create a nickname based on the first letter of the
+ * first name and the last name.
+ *
+ * @param {string} nickname - The nickname of the user.
  * @param {string} firstName - The first name of the user.
  * @param {string} lastName - The last name of the user.
- * @return {string} The created nickname
+ * @return {string} The normalized nickname.
  */
-function createNickname(firstName: string, lastName: string): string {
-  return (firstName[0] + lastName) // first letter of first name + last name
-    .toLowerCase()
-    .normalize("NFD") // remove accents
-    .replace(/\p{Diacritic}/gu, ""); // remove diacritics
+function createNickname(nickname: string, firstName: string, lastName: string): string {
+  const n = nickname || firstName[0] + lastName; // first letter of first name + last name
+  return n.toLowerCase()
+          .normalize("NFD") // remove accents
+          .replace(/\p{Diacritic}/gu, ""); // remove diacritics
 }
 
 export {