+++ tests/normalizeName.test.js
@@ -18,23 +18,33 @@ test("sends true to normalizeName", () => {
 });
 
 test("sends empty string to normalizeName", () => {
-  expect(normalizeName("")).toBe(true);
-});
-
-test("sends empty string to normalizeName with empty check", () => {
-  expect(normalizeName("", true)).toBe(false);
+  expect(normalizeName("")).toBe(false);
 });
 
 test("sends false to normalizeName", () => {
   expect(normalizeName(false)).toBe(false);
 });
 
-test("sends string to normalizeName", () => {
-  expect(normalizeName("string")).toBe(true);
+const str = "string";
+const strRes = "String";
+
+test("sends string to ucfirst", () => {
+  expect(normalizeName(str)).toBe(strRes);
 });
 
-test("sends string to normalizeName with empty check", () => {
-  expect(normalizeName("string", true)).toBe(true);
+test("sends string to ucfirst without everyWords", () => {
+  expect(normalizeName(str, false)).toBe(strRes);
+});
+
+const strstr = "string string";
+const strstrRes = "String String";
+
+test("sends string string to ucfirst", () => {
+  expect(normalizeName(strstr, true)).toBe(strstrRes);
+});
+
+test("sends string string to ucfirst without everyWords", () => {
+  expect(normalizeName(strstr, false)).toBe(strstrRes);
 });
 
 test("sends positive even integer to normalizeName", () => {
@@ -99,38 +109,18 @@ test("sends array to normalizeName with empty check", () => {
 
 var json = `{
   "actor": {
-    "name": "Tom Cruise",
-    "age": 56,
-    "Born At": "Syracuse, NY",
-    "Birthdate": "July 3 1962",
-    "photo": "https://jsonformatter.org/img/tom-cruise.jpg"
+    "name": "Tom Cruise"
   }
 }`;
 
-test("sends json to normalizeName", () => {
-  expect(normalizeName(json)).toBe(true);
-});
-
-test("sends json to normalizeName with empty check", () => {
-  expect(normalizeName(json, true)).toBe(true);
-});
-
-var invalidjson = `{
-  "actor: {
-    "name": "Tom Cruise",
-    "age": 56
-    "Born At": "Syracuse, NY",
-    "Birthdate": "July 3 1962",
-    "photo": "https://jsonformatter.org/img/tom-cruise.jpg"
+var jsonRes = `{
+  "actor": {
+    "name": "tom Cruise"
   }
 }`;
 
-test("sends invalid json to normalizeName", () => {
-  expect(normalizeName(invalidjson)).toBe(true);
-});
-
-test("sends invalid json to normalizeName with empty check", () => {
-  expect(normalizeName(invalidjson, true)).toBe(true);
+test("sends json string to normalizeName", () => {
+  expect(normalizeName(json)).toBe(jsonRes);
 });
 
 function testFunction() {